Trekkjer ut EtterlatteBrevKode i eiga fil, for tydeleggjering
