+++ pensjon-brevbaker/src/main/kotlin/no/nav/pensjon/etterlatte/EtterlatteBrevRequest.kt
@@ -3,64 +3,4 @@ package no.nav.pensjon.etterlatte
 import no.nav.pensjon.brevbaker.api.model.Felles
 import no.nav.pensjon.brevbaker.api.model.LanguageCode
 
-enum class EtterlatteBrevKode {
-    AVVIST_KLAGE_INNHOLD,
-    AVVIST_KLAGE_FERDIG,
-    BARNEPENSJON_AVSLAG,
-    BARNEPENSJON_AVSLAG_UTFALL,
-    BARNEPENSJON_INNVILGELSE,
-    BARNEPENSJON_INNVILGELSE_UTFALL,
-    BARNEPENSJON_INNVILGELSE_FORELDRELOES,
-    BARNEPENSJON_INNVILGELSE_UTFALL_FORELDRELOES,
-    BARNEPENSJON_OPPHOER,
-    BARNEPENSJON_OPPHOER_UTFALL,
-    BARNEPENSJON_REVURDERING,
-    BARNEPENSJON_REVURDERING_UTFALL,
-    BARNEPENSJON_VARSEL,
-    BARNEPENSJON_VARSEL_UTFALL,
-    BARNEPENSJON_VEDTAK_OMREGNING,
-    BARNEPENSJON_VEDTAK_OMREGNING_FERDIG,
-    BARNEPENSJON_FORHAANDSVARSEL_OMREGNING,
-    BARNEPENSJON_VEDLEGG_BEREGNING_TRYGDETID_UTFALL,
-    BARNEPENSJON_VEDLEGG_FORHAANDSVARSEL_UTFALL,
-    BARNEPENSJON_INFORMASJON_DOEDSFALL,
-    BARNEPENSJON_INFORMASJON_DOEDSFALL_MELLOM_ATTEN_OG_TJUE_VED_REFORMTIDSPUNKT,
-    BARNEPENSJON_INFORMASJON_MOTTATT_SOEKNAD,
-    BARNEPENSJON_INFORMASJON_INNHENTING_AV_OPPLYSNINGER,
-    OMSTILLINGSSTOENAD_INFORMASJON_DOEDSFALL,
-    OMSTILLINGSSTOENAD_INFORMASJON_MOTTATT_SOEKNAD,
-    OMSTILLINGSSTOENAD_INFORMASJON_INNHENTING_AV_OPPLYSNINGER,
-    OMSTILLINGSSTOENAD_AKTIVITETSPLIKT_VARSELBREV_INNHOLD,
-    OMSTILLINGSSTOENAD_AKTIVITETSPLIKT_INFORMASJON_4MND_INNHOLD,
-    OMSTILLINGSSTOENAD_AKTIVITETSPLIKT_INFORMASJON_6MND_INNHOLD,
-    OMSTILLINGSSTOENAD_AKTIVITETSPLIKT_INFORMASJON_10MND_INNHOLD,
-    OMSTILLINGSSTOENAD_AVSLAG,
-    OMSTILLINGSSTOENAD_AVSLAG_UTFALL,
-    OMSTILLINGSSTOENAD_INNVILGELSE,
-    OMSTILLINGSSTOENAD_INNVILGELSE_UTFALL,
-    OMSTILLINGSSTOENAD_OPPHOER,
-    OMSTILLINGSSTOENAD_OPPHOER_UTFALL,
-    OMSTILLINGSSTOENAD_REVURDERING,
-    OMSTILLINGSSTOENAD_REVURDERING_UTFALL,
-    OMSTILLINGSSTOENAD_VARSEL,
-    OMSTILLINGSSTOENAD_VARSEL_UTFALL,
-    OMSTILLINGSSTOENAD_VARSEL_AKTIVITETSPLIKT,
-    OMSTILLINGSSTOENAD_VARSEL_AKTIVITETSPLIKT_UTFALL,
-    OMSTILLINGSSTOENAD_VEDLEGG_BEREGNING_UTFALL,
-    OMSTILLINGSSTOENAD_VEDLEGG_FORHAANDSVARSEL_UTFALL,
-    OMSTILLINGSSTOENAD_INNTEKTSJUSTERING_VARSEL,
-    OMSTILLINGSSTOENAD_INNTEKTSJUSTERING_VEDTAK,
-    OMSTILLINGSSTOENAD_INNTEKTSJUSTERING_VEDTAK_UTFALL,
-    TILBAKEKREVING_INNHOLD,
-    TILBAKEKREVING_FERDIG,
-    TOM_MAL,
-    TOM_DELMAL,
-    TOM_MAL_INFORMASJONSBREV,
-    UTSATT_KLAGEFRIST,
-    KLAGE_OVERSENDELSE_BLANKETT,
-    KLAGE_OVERSENDELSE_BRUKER,
-    KLAGE_SAKSBEHANDLINGS_INFO,
-    INGEN_REDIGERBAR_DEL, // Til bruk for å modellere at brevet ikke har en redigerbar del
-}
-
 data class EtterlatteBrevRequest(val kode: EtterlatteBrevKode, val letterData: Any, val felles: Felles, val language: LanguageCode)
